MQTT/json_parser/requirements.txt
python-dotenv
paho-mqtt
requests